requirements.txt
django

django-debug-toolbar

psycopg2-binary

python-binance
